+++ PersonalBest/Model/Record/RecordEntry.swift
@@ -5,7 +5,7 @@ import Foundation
 protocol PRecordEntry {
     var date: Date { get }
     var variantName: String? { get }
-    var entryValues: [MeasurementType: Decimal] { get }
+    var entryValues: [MeasurementType: Double] { get }
 }
 
 /// record entry not linked to core data
@@ -13,6 +13,6 @@ struct RecordEntry: PRecordEntry {
     
     let date: Date
     let variantName: String?
-    let entryValues: [MeasurementType: Decimal]
+    let entryValues: [MeasurementType: Double]
     
 }
